refactor: 미사용 이미지 삭제 Batch 관련 클래스를 unusedimagebatch 패키지로 분리
